requirements.txt
biopython
numpy<2
pandas
scikit-learn
tqdm
joblib
matplotlib
seaborn
